% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.blacken_NA_util}
\alias{.blacken_NA_util}
\title{.blacken_NA_util}
\usage{
.blacken_NA_util(x_list)
}
\arguments{
\item{x}{A list of raster objects}
}
\value{
a raster object with NAs replaced by 0
lapply(r_list_out_stretched,FUN = rtsVis:::.blacken_NAs
}
\description{
.blacken_NA_util
}
