% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.ts_guess_raster_type}
\alias{.ts_guess_raster_type}
\title{ts_guess_raster_type}
\usage{
.ts_guess_raster_type(x)
}
\arguments{
\item{x}{A raster object}
}
\value{

}
\description{
Attempts to determine the raster type of the input raster, to be used for other functions in the rtsVis package.
}
