% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.ts_makeframes}
\alias{.ts_makeframes}
\title{.ts_makeframes}
\usage{
.ts_makeframes(x_list, r_type = "RGB")
}
\arguments{
\item{x_list}{a list of rasters}

\item{r_type}{one of "discrete","gradient", "RGB"}
}
\value{
a list of ggplots, carrying over the "time" attribute of x_list set
}
\description{
.ts_makeframes
}
