% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.ts_update_NA_util}
\alias{.ts_update_NA_util}
\title{.ts_update_NA_util}
\usage{
.ts_update_NA_util(x, new_na)
}
\arguments{
\item{x}{a raster object}

\item{new_na}{Value to replace the old one as NA}
}
\value{
The modified raster object, with the new NA value set
}
\description{
.ts_update_NA_util
}
