% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MODIS_SI_ds}
\alias{MODIS_SI_ds}
\title{Example MODIS time series for southern Slovenia.}
\format{
A list of 105 raster objects
}
\source{
\url{https://lpdaac.usgs.gov/tools/usgs-earthexplorer/}
}
\usage{
MODIS_SI_ds
}
\description{
105 MODIS MCD43A4 Images covering the time from 2020-02-15 to 2020-05-29 ,
downsampled by a factor of 4 and subset to bands 1 to 4
to serve as lightweight example time series.
Acquired from LP DAAC using the getSpatialData package.
}
\keyword{datasets}
