% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SI_positions}
\alias{SI_positions}
\title{Example vector data for Slovenia.}
\format{
A list of three sets of example vector data:
 \describe{
  \item{points_matrix}{A matrix of coordinates corresponding to the coordinates of two points in Slovenia. Manually created. MODIS sinusoidal projection.}
  \item{points}{SpatialPointsDataFrame of four points in Slovenian municipalities. Manually created. MODIS sinusoidal projection.}
  \item{polygons}{SpatialPolygonsDataFrame of three polygons corresponding to borders of municipalities. Acquired from GADM. MODIS sinusoidal projection.}
  ...
}
}
\source{
\url{https://gadm.org/data.html}
}
\usage{
SI_positions
}
\description{
Three sets of example vector data for Slovenia. Two sp objects and one matrix object. Chosen to match
the "MODIS_SI_ds" raster data.
}
\keyword{datasets}
