% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_add_positions_to_frames.R
\name{ts_add_positions_to_frames}
\alias{ts_add_positions_to_frames}
\title{Add points, coordinates, or polygons to a list of spatial plotss}
\usage{
ts_add_positions_to_frames(
  r_frame_list,
  positions,
  position_names = NULL,
  pcol = "red",
  psize = 2,
  position_legend_title = "Position",
  legend_position = "right"
)
}
\arguments{
\item{r_frame_list}{}

\item{positions}{}

\item{position_names}{}
}
\value{

}
\description{
Add points, coordinates, or polygons to a list of spatial plotss
}
