% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_frametimes.R
\name{ts_copy_frametimes}
\alias{ts_copy_frametimes}
\title{Sets frametimes for a list of frames}
\usage{
ts_copy_frametimes(x, y)
}
\arguments{
\item{x}{A list to which timestamps are added.}

\item{y}{A list from which the timestamps are derived.}
}
\value{
A list identical to x with timestamps carried over from y.
}
\description{
Sets frametimes for a list of frames
}
