% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ts_get_layer_quantiles}
\alias{ts_get_layer_quantiles}
\title{ts_get_layer_quantiles}
\usage{
ts_get_layer_quantiles(x, minq = 0.02, maxq = 0.98, samplesize = 1e+05)
}
\arguments{
\item{x}{a raster layer}

\item{minq}{Lower quantile to be determined}

\item{maxq}{Upper quantile to be determined}

\item{samplesize}{Number of samples to take}
}
\value{
a named vector of the minimum and maximum value (corresponding to minq and maxq)
}
\description{
Takes regular samples from a single rasterlayer and determines the quantiles
}
