% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ts_get_stack_quantiles}
\alias{ts_get_stack_quantiles}
\title{ts_get_stack_quantiles}
\usage{
ts_get_stack_quantiles(x, minq = 0.02, maxq = 0.98, samplesize = 1e+05)
}
\arguments{
\item{x}{a raster stack}

\item{minq}{Lower quantile to be determined}

\item{maxq}{Upper quantile to be determined}

\item{samplesize}{Number of samples to take}
}
\value{
a matrix of n_layers*2, one min and one max bound for every layer
}
\description{
Takes regular samples from a single rasterstack and determines the quantiles for every layer
}
