% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ts_get_ts_quantiles}
\alias{ts_get_ts_quantiles}
\title{ts_get_ts_quantiles}
\usage{
ts_get_ts_quantiles(ts, minq = 0.02, maxq = 0.98, samplesize = 1e+05)
}
\arguments{
\item{ts}{a list of raster stacks}

\item{minq}{Lower quantile to be determined}

\item{maxq}{Upper quantile to be determined}

\item{samplesize}{Number of samples to take}
}
\value{
a matrix of n_layers*2, one min and one max bound for every layer, which is the min and the max respectively of all the rasters for that layer index
}
\description{
ts_get_ts_quantiles
}
