% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_makeframes.R
\name{ts_makeframes}
\alias{ts_makeframes}
\title{ts_makeframes}
\usage{
ts_makeframes(
  x_list,
  r_type = NULL,
  minq = 0.02,
  maxq = 0.98,
  samplesize = 1000,
  blacken_NA = F,
  l_indices = NULL
)
}
\arguments{
\item{x_list}{A list of rasters.}

\item{r_type}{one of "discrete","gradient", "RGB"}

\item{minq}{Lower quantile boundary for the stretch}

\item{maxq}{Upper quantile boundary for the stretch}

\item{samplesize}{Number of samples to determine the quantile from}
}
\value{
A list of ggplots
}
\description{
ts_makeframes
}
