% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ts_stretch}
\alias{ts_stretch}
\title{ts_stretch}
\usage{
ts_stretch(x, minqs, maxqs, ymin = 0, ymax = 0)
}
\arguments{
\item{x}{a raster to be stretched}

\item{minqs}{Lower quantile}

\item{maxqs}{Higher quantile}

\item{ymin}{target min value}

\item{ymax}{target max value}
}
\value{
A raster with values between ymin and ymax
}
\description{
Stretch and clips a raster(stack) from within a certain source range determined by min quantile and max quantile to a target range ymin to ymax
}
