% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ts_stretch_list}
\alias{ts_stretch_list}
\title{ts_stretch_list}
\usage{
ts_stretch_list(
  x_list,
  minq = 0.01,
  maxq = 0.99,
  ymin = 0,
  ymax = 0,
  samplesize = 10000
)
}
\arguments{
\item{x_list}{a list of rasters to be stretched}

\item{ymin}{target min value}

\item{ymax}{target max value}

\item{minqs}{Lower quantile}

\item{maxqs}{Higher quantile}
}
\value{
A list of raster(stacks) with values between ymin and ymax
}
\description{
Stretch and clips a list of raster(stacks) from within a certain source range determined by min quantile and max quantile to a target range ymin to ymax. The quantiles are determined for each layer individually but across all rasters of the inputlist
}
