% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_add_positions_to_frames.R
\name{ts_add_positions_to_frames}
\alias{ts_add_positions_to_frames}
\title{Add points, coordinates, or polygons to a list of spatial plots}
\usage{
ts_add_positions_to_frames(
  r_frame_list,
  positions,
  position_names = NULL,
  pcol = "red",
  tcol = "red",
  psize = 2,
  tsize = 7,
  ttype = "text",
  t_hjust = 0,
  t_vjust = 0,
  position_legend_title = "Position",
  legend_position = "right",
  aes_by_pos = F,
  add_text = F
)
}
\arguments{
\item{r_frame_list}{list of ggplots, as generated by \code{\link{ts_makeframes}}.}

\item{positions}{object containing the coordinates. One of \itemize{
 \item{A two-column \code{matrix} of coordinates where the first column corresponds to the longitude and the second column corresponds to the latitude.}
 \item{A \code{SpatialPolygonsDataFrame} from which \link[sp]{coordinates} can be extracted.}
 \item{A \code{SpatialPointsDataFrame} from which \link[sp]{coordinates} can be extracted.}
}}

\item{position_names}{(Optional) character, names of the positions to be added in legend or text (If \code{add_text}). By default, will create placeholder names by combining the object type and Id (Example: *"Polygon 3"*)}

\item{pcol}{(Optional) character, color of the spatial objects. Default is \code{"red"}.}

\item{tcol}{(Optional) character, if \code{add_text}: The color of the text. Default is \code{"red"}.}

\item{psize}{(Optional) numeric, plot size of the spatial objects. Default is \code{2}.}

\item{tsize}{(Optional) numeric, if \code{add_text}: The size of the text. Default is \code{7}.}

\item{ttype}{(Optional) character, if \code{add_text}: The type of the text. Either \code{"label"} or \code{"text"}. Default is \code{"text"}.}

\item{t_hjust}{(Optional) numeric, if \code{add_text}: Horizontal offset of the text in map units. Default is \code{0}.}

\item{t_vjust}{(Optional) numeric, if \code{add_text}: Vertical offset of the text in map units. Default is \code{0}.}

\item{position_legend_title}{(Optional) character, title of the legend. Default is \code{"Position"}.}

\item{legend_position}{(Optional) character, position of the legend. Use \code{"none"} to disable the legend. Default is \code{"right"}}

\item{aes_by_pos}{(Optional) logical. If \code{TRUE}: vary some aesthetic (linetype for polygons, shape for points) to be different for each position? If  \code{FALSE}, this also disables the legend, as no notable classes will be plotted. Default is \code{FALSE}.}

\item{add_text}{(Optional) logical.  If \code{TRUE}: add a text to each position using \link[moveVis]{add_text}.  Default is \code{"False"}.}
}
\value{
A list of ggplots with added positions.
}
\description{
Add points, coordinates, or polygons to a list of spatial plots
}
\details{
The function takes a \code{positions} object, which can be a spatial object or a matrix of coordinates, and adds them to each of the elements of \code{r_frame_list} using \link[moveVis]{add_gg}. Optionally it also adds text at their respective positions using\link[moveVis]{add_text}.

- \code{ts_add_positions_to_frames} is intended to be an easy way to add multiple objects to the spatial frames at fixed positions. 
For adding individual positions or text, potentially at varying positions, it is recommended to all \link[moveVis]{add_gg} and \link[moveVis]{add_text} directly.
}
\seealso{
\link[moveVis]{add_text} \link[moveVis]{add_gg}
}
\author{
Johannes Mast
}
