% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_flow_frames.R
\name{ts_flow_frames}
\alias{ts_flow_frames}
\title{Create line charts of a raster time series}
\usage{
ts_flow_frames(
  r_list,
  positions,
  position_names = NULL,
  val_min = NULL,
  val_max = NULL,
  val_by = 0.1,
  path_size = 1,
  path_legend = T,
  path_legend_title = "Positions"
)
}
\arguments{
\item{r_list}{}

\item{positions}{}

\item{position_names}{}

\item{val_min}{}

\item{val_max}{}

\item{val_by}{}

\item{path_size}{}

\item{path_legend}{}

\item{path_legend_title}{}
}
\value{

}
\description{
Create line charts of a raster time series
}
