% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_raster.R
\name{ts_raster}
\alias{ts_raster}
\title{Assemble/interpolate a raster time series}
\usage{
ts_raster(r_list, r_times, out_times = NA, fade_raster = F, ..., verbose = T)
}
\arguments{
\item{r_list}{list of rasters}

\item{r_times}{POSIXct,}

\item{out_times}{POSIXct,}

\item{fade_raster}{logical,}

\item{...}{additional arguments.}

\item{verbose}{}
}
\value{

}
\description{
This function assembles a raster time series by assigning or interpolating input rasters to a target time series.
}
